requirements.txt
elasticsearch==7.17.6
elasticsearch_dsl==7.4.0
frontend==0.0.3
numpy==1.21.5
pandas==1.4.4
sentence_transformers==2.2.2
spacy==3.4.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
python-dotenv
blinker
gunicorn==20.1.0
Authlib==1.2.0
psycopg2-binary
requests==2.28.1
werkzeug==2.0.3
markupsafe==2.0.1
